Extract SniSmuxHeader to its own file
